Hide some internal things in an Internal module
